==> main_study/code/00_preprocessData.ipynb <==
import re
import json
import random
import pandas as pd

#!wget https://raw.githubusercontent.com/Instruction-Tuning-with-GPT-4/GPT-4-LLM/main/data/alpaca_gpt4_data.json

dataset_file = "alpaca_gpt4_data.json"

with open(dataset_file, "r") as f:
    alpaca = json.load(f)

seed = 42

random.seed(seed)
random.shuffle(alpaca)
type(alpaca), alpaca[0:3], len(alpaca)

# filter for elements in the list where input key is empty
filtered_alpaca = [item for item in alpaca if item['input'] == '']
type(filtered_alpaca), filtered_alpaca[0:3], len(filtered_alpaca)

#remove refusals
refusal_pattern = r"\b(?:I will not|I(?:'m| am) unable to|I cannot|I can't|I'm sorry|I am sorry|I am not able|I'm unable|I'm not able|I am unable|I am unable |AI assistant|AI chatbot| AI language model|As an AI)\b"
filtered_alpaca = [item for item in filtered_alpaca if not re.search(refusal_pattern, item['output'], re.IGNORECASE)]
type(filtered_alpaca), filtered_alpaca[0:3], len(filtered_alpaca)

# split
train_dataset = filtered_alpaca[:10000]
eval_dataset = filtered_alpaca[-1000:]

train_df = pd.DataFrame(train_dataset)
eval_df = pd.DataFrame(eval_dataset)

train_df.to_json("GPT-4_Alpaca_train.jsonl", orient='records', lines=True)
eval_df.to_json("GPT-4_Alpaca_eval.jsonl", orient='records', lines=True)
